style: refactor serializers from runner
