+++ backend/api/models/courseModel.py
@@ -1,6 +1,7 @@
 from django.contrib.auth.models import User
 from django.db import models
 from django.utils.translation import gettext_lazy as _
+from .studentGroupModel import StudentGroup
 
 
 class Course(models.Model):
@@ -119,30 +120,3 @@ class Course(models.Model):
         # Save groups
         for group in student_groups:
             group.save()
-
-
-class StudentGroup(models.Model):
-    """
-    A group of students, related to one course.
-    """
-
-    course = models.ForeignKey(Course, on_delete=models.CASCADE)
-    name = models.CharField(max_length=255)
-    students = models.ManyToManyField(
-        User, related_name="%(class)s_students", blank=True
-    )
-
-    def save(self, *args, **kwargs):
-        is_new = self._state.adding
-
-        super(StudentGroup, self).save(*args, **kwargs)
-
-        if is_new:
-            self.course.handle_student_groups_change()
-
-    def delete(self, *args, **kwargs):
-        super(StudentGroup, self).delete(*args, **kwargs)
-        self.course.handle_student_groups_change()
-
-    def __str__(self):
-        return f"{self.name} ({self.course.title})"
