+++ backend/api/models/submissionModel.py
@@ -3,6 +3,8 @@ from celery import Celery
 from django.contrib.auth.models import User
 from django.db import models
 from django.utils.translation import gettext_lazy as _
+from .testResultModel import TestResult
+from .testModel import Test
 
 
 class Submission(models.Model):
