+++ backend/api/serializers/__init__.py
@@ -3,3 +3,6 @@ from .courseSerializers import *  # noqa: W0611
 from .exerciseSerializers import *  # noqa: W0611
 from .sessionSerializers import *  # noqa: W0611
 from .studentGroupSerializers import *  # noqa: W0611
+from .submissionSerializers import *  # noqa: W0611
+from .testResultSerializers import *  # noqa: W0611
+from .testSerializers import *  # noqa: W0611
