+++ backend/api/serializers/testResultSerializers.py
@@ -3,24 +3,6 @@ from rest_framework import serializers, validators
 import copy
 
 
-class TestSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Test
-        fields = ["id", "exercise", "name", "stdin", "stdout", "coefficient"]
-
-
-class SubmissionSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Submission
-        fields = ["id", "exercise", "file", "status", "created"]
-
-    def run_validators(self, value):
-        for validator in copy.copy(self.validators):
-            if isinstance(validator, validators.UniqueTogetherValidator):
-                self.validators.remove(validator)
-        super(SubmissionSerializer, self).run_validators(value)
-
-
 class TestResultSerializer(serializers.ModelSerializer):
     # Fixes depth=1 ignoring fields
     submission_pk = serializers.PrimaryKeyRelatedField(
