+++ backend/api/serializers/testSerializers.py
@@ -1,0 +1,8 @@
+from api.models import Test
+from rest_framework import serializers
+
+
+class TestSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Test
+        fields = ["id", "exercise", "name", "stdin", "stdout", "coefficient"]
